Move audio recorder logic into its own file
